Support vpnkit <-> hyperkit protocol
It removes the need of vsock and the client running inside the VM.
Hyperkit directly creates a virtual network card connected the daemon.

Hyperkit doesn't support a MTU larger than 1500.

Hyperkit and Qemu use almost the same protocol. The only difference is
the byte order and the size field length.