Add Qemu support
User can now choose between vpnkit or qemu protocol.

Qemu needs to be run with this argument:
`-netdev socket,id=vlan,connect=IP:PORT -device virtio-net-pci,netdev=vlan`
where IP and PORT represent the daemon on the host.

This extra port allocation can be avoided with the qemu-wrapper in cmd/.
The daemon can listen a unix domain socket, the wrapper will pass it to
Qemu as a file descriptor.