Run defers on exit and add cmd line opts
Slightly altered the structure for the host binary so that when the
process is killed or control-c'd, it will exit gracefully by running all
the defer statements which generally clean up the sockets and pidfiles.

Also added command line options for --pid-file, which stores the pid so
it can be looked up later, as well as --ssh-port that allows the ssh
port to be dynamic (default is 2222).

Finally, removed the secondary http server on port 8080.

Signed-off-by: Brent Baude <bbaude@redhat.com>