connect over stdio for wsl
This adds stdio as a way to communicate between gvproxy and vm mainly for use
with WSL2, although it should work for other cases as well.

When network connections between WSL2 and the Windows host are blocked, stdio
is the only reliable way to establish a channel between WSL2 and the Windows
host. Hyper-V socket for WSL2 is a possibility, but it requires undocumented
APIs and admin privileges.

Signed-off-by: Keiichi Shimamura <sakai135@users.noreply.github.com>