Document Qemu 7.2.0 QAPI changes
Starting from version 7.2.0 there is no need to use `qemu-wrapper`.

New options described here:
https://github.com/qemu/qemu/blob/7c9236d6d61f30583d5d860097d88dbf0fe487bf/qemu-options.hx#L2773

Now it is also possible to use Unix domain socket on Windows OS (as FD
approach is not supported on that OS), but as there was no Windows
documentation prior to this moment (even for TCP sockets) it is not
provided.

Signed-off-by: Arthur Sengileyev <arthur.sengileyev@gmail.com>