hvsock related systemd and nm config files
the network manager file configures a vsock0 network interface with the
specific mac-address that gvproxy needs.  this must be run before the vm
command is run.

the systemd file is for running the vm command after network manager
creates the network interface.  the url must be modified with the vm ID
when using hvsock and Microsoft HyperV virtualization.  Once the vm
command has been run, we issue a postexec to ask NM to bring the "link"
up.  this may not be entirely necessary but wfm.

Signed-off-by: Brent Baude <bbaude@redhat.com>