service/dns: Compress dns message
Recently we observed, dns messages from our dns service is not compress
which sometime makes message size more than 512B and some client tools
fails to process it.

In this PR, message compression is used to reduce the length of DNS
messages by removing duplicated information.

- https://spathis.medium.com/how-dns-got-its-messages-on-diet-c49568b234a2

Without this PR if we query `dig
secure-feeds-production-us-east-1-761931097553.s3.us-east-1.amazonaws.com`
then message size is 803B and with this PR now it is 242B.

- Issue: https://github.com/crc-org/crc/issues/3643

Signed-off-by: Praveen Kumar <kumarpraveen.nitdgp@gmail.com>