Add listen to BSDs
Copy over Darwin's code and reuse it for the most common BSDs. It is
untested in all BSDs added but in DragonFly BSD but at least
it'll allow the build to succeed.

TODO: Might want to merge darwin and BSD file together but for now
they are separate because we might need to add something BSD-specific.