Fix golangci-lint errors
Fixes golangci-lint errors introduced by #258

If `gosec` linting is enabled for the return statement in `Cmd`, an
error will be returned: `G204: Subprocess launched with a potential tainted
input or cmd arguments (gosec)`. This error tries to make sure a
user cannot provide a binary that could cause harm to the system.
We can't do much about this since the binary is externally
provided. It is up to the caller to make sure they don't provide a
"dangerous" binary. Therefore, it's ok that we ignore this linting
error.

Signed-off-by: Jake Correnti <jakecorrenti+github@proton.me>