Add support for volume sharing on gvproxy for Windows
Specifically, gvproxy will act as a 9p server, serving shares
requested by the caller on HyperV vsocks, to allow the guest VM
to access content on the host.

The vsocks are intended to be managed by the caller in this
model. As such, gvproxy receives the path to be shared and a
vsock port number to share it on via CLI. An arbitrary number of
these are accepted, as each share needs a separate server and
vsock (they will be mounted by the Linux kernel 9p code within
the guest VM, which does not support multiplexing multiple shares
over a single vsock).

Also, make a slight change to .gitignore - the ./bin/ ignore
pattern was not ignoring the bin/ directory for me (removing the
leading . fixes things).

Signed-off-by: Matthew Heon <matthew.heon@pm.me>