version: Add versioning for github tarballs
When building from github tarballs, getting the version from `git
describe` won't work, or worse, will be irrelevant [1]
However, we can make use of $Format$ and .gitattributes as described in
[2] to automatically substitute the correct version in a go variable
when `git archive` is used.
In particular, the correct version number will automatically be
substituted when GitHub creates release tarballs.

[1] rpms/gvisor-tap-vsock.spec uses `%autosetup -Sgit -n %{name}-%{version}` which unpacks
the release tarball in a newly created git repository.

[2] https://icinga.com/blog/2022/05/25/embedding-git-commit-information-in-go-binaries/

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>