ssh: Recreate connection on retries in setupProxy
The previous fix was not working as expected, as the ssh go code will
close the underlying connection when there's a failure.
This was causing the retries for CreateBastion() to fail, as after the
first failure it would try to use a closed connection.

This commit recreates the connection each time before calling
CreateBastion() to fix this. This also simplifies the code.

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>