ghactions: setup-go: Use go version from go.mod
This commit changes uses of actions/setup-go to pick the go version
required in go.mod.
This means we'll be using the same go version everywhere, and since
there is no major version hardcoded (such as '1.20.x'), this means we
don't need to remember to update it once in a while.

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>