handle ENOBUFS when writing to VM socket
When a high throughput is happeing we send a lot of data into the VM
socket, the VM has to read it in order to empty the buffer. This is
inherently race so it is possible to get ENOBUFS here. In this case just
keep trying writing until it works.

Fixes #367

Signed-off-by: Paul Holzinger <pholzing@redhat.com>