test: Use crc.dev records for DNS tests
test/basic_test.go has some DNS resolution tests for various types of
DNS records. These tests use wikipedia.org or google.com, which means
they have to be updated once in a while when the admins of these sites
decide to change these records.
Lately this has happened a few times with wikipedia SRV and MX record.
This commit switches to crc.dev for these tests, as it's a domain we
control, and thus we have more control over DNS record changes.

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>